move hbx admin controller under individual market namespace
