+++ app/assets/javascripts/aptc.js.erb
@@ -28,7 +28,7 @@ var APTCModule = ( function( window, undefined ) {
     $.ajax({
       type: "GET",
       data:{person_id: person_id, family_id: family_id, max_aptc: max_aptc, csr_percentage: csr_percentage, applied_aptcs_array: applied_aptcs_array, year: year},
-      url: "/hbx_admin/calculate_aptc_csr",
+      url: "individual_market/hbx_admin/calculate_aptc_csr",
       success: function (d) {
           //$('#loading').html("");
       }
@@ -44,7 +44,7 @@ var APTCModule = ( function( window, undefined ) {
     $.ajax({
       type: "GET",
       data:{person_id: person_id, family_id: family_id, year: year},
-      url: "/hbx_admin/edit_aptc_csr",
+      url: "individual_market/hbx_admin/edit_aptc_csr",
       success: function (d) {
         $('#loading').html("");
       }
@@ -60,7 +60,7 @@ var APTCModule = ( function( window, undefined ) {
     $.ajax({
       type: "GET",
       data:{person_id: person_id, family_id: family_id, year: year},
-      url: "/hbx_admin/edit_aptc_csr",
+      url: "individual_market/hbx_admin/edit_aptc_csr",
       success: function (d) {
               $('#loading').html("");
       }
