+++ config/routes.rb
@@ -15,9 +15,9 @@ Rails.application.routes.draw do
   get 'unsupportive_browser' => 'users#unsupportive_browser'
 
   match "hbx_admin/about_us" => "hbx_admin#about_us", as: :about_us, via: :get
-  match "hbx_admin/update_aptc_csr" => "hbx_admin#update_aptc_csr", as: :update_aptc_csr, via: [:get, :post]
-  match "hbx_admin/edit_aptc_csr" => "hbx_admin#edit_aptc_csr", as: :edit_aptc_csr, via: [:get, :post], defaults: { format: 'js' }
-  match "hbx_admin/calculate_aptc_csr" => "hbx_admin#calculate_aptc_csr", as: :calculate_aptc_csr, via: :get
+  match "individual_market/hbx_admin/update_aptc_csr" => "individual_market/hbx_admin#update_aptc_csr", as: :update_aptc_csr, via: [:get, :post]
+  match "individual_market/hbx_admin/edit_aptc_csr" => "individual_market/hbx_admin#edit_aptc_csr", as: :edit_aptc_csr, via: [:get, :post], defaults: { format: 'js' }
+  match "individual_market/hbx_admin/calculate_aptc_csr" => "individual_market/hbx_admin#calculate_aptc_csr", as: :calculate_aptc_csr, via: :get
   post 'show_hints' => 'welcome#show_hints', :constraints => { :only_ajax => true }
 
   post 'submit_notice' => "hbx_admin#submit_notice", as: :submit_notice
